# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Requirements Platform API",
        "description": "A comprehensive API for the Requirements Gathering Platform with AI Interviews, Project Management, and Vendor Marketplace",
        "contact": {
            "email": "api@requirementsplatform.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://specfirst-api.netsolutionindia.com",
            "description": "Staging Environment"
        },
        {
            "url": "https://api.yourapp.com",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/project-types/{slug}": {
            "get": {
                "tags": [
                    "Project Types"
                ],
                "summary": "Get project type by slug",
                "description": "Retrieve a specific project type by its slug",
                "operationId": "b6d2d9cef20f185e45abf23d9c33055e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "web_app"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project type retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Web App"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "web_app"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Dashboards, SaaS, Portals"
                                                },
                                                "icon": {
                                                    "type": "string",
                                                    "example": "🌐"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#3b82f6"
                                                },
                                                "sort_order": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project type not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Project type not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}